Split ParserState into abstract and BaseParserState
The trig ParserState is now abstract, getting around strict typing
semantics enforced by e.g. the MyPy compiler (mypyc) when passing None
to a BaseParserState (as done for the top level). This workaround now
uses an instance of the abstract ParserState instead (dubious enough but
working since it is not accessed at the top level).