Transpile docloader by handling Protocol calls
If a function definition matches a Protocol class, track that, and if
target language defines a protocol_call, use that as method to call in
transpiled code.

This includes declaring transpiled functions as anonymous classes
implementing the Protocol, which in Java is transpiled to an interface.

Declare BiFunction as the implemented interface of LoadDocumentCallback.